Rename ZombieGazeEffect.h/cpp to AumtlichGazeEffect.h/cpp
